# perf: avoid unnecessary overrides