# Pull Request Check List

Related to #9956

<!-- This is just a reminder about the most common mistakes. Please make sure that you tick all *appropriate* boxes.  But please read our [contribution guide](https://python-poetry.org/docs/contributing/) at least once, it will save you unnecessary review cycles! -->

- [ ] Added **tests** for changed code.
- [ ] Updated **documentation** for changed code.

If there is only one or no relevant requirement among the duplicates after filtering, it is not necessary to raise an `OverrideNeededError`. By moving the filtering upwards a little and checking if there are fewer than 2 duplicates left, we can avoid unnecessary overrides.

Testing with the example from https://github.com/python-poetry/poetry/issues/9956#issuecomment-2614554740:

poetry version|lock (--regenerate)|lock (--no-update)
---|---|---
1.8|1040 s|107 s
without PR<sup>1</sup>|>1400 s|255 s
with PR<sup>1</sup>|441 s|35 s

<sup>1</sup> including python-poetry/poetry-core#832 to fix the endless recursion mentioned in https://github.com/python-poetry/poetry/issues/9956#issuecomment-2629031927

## Summary by Sourcery

Enhancements:
- Optimize dependency resolution by filtering irrelevant requirements before checking for overrides.